Molfino
Strip Light
A LED strip light (which is also known as LED Tape or Ribbon Light) is a flexible circuit board populated by Surface Mounted light and Light Emitting Diodes (SMD LED) and many other components that usually come with an adhesive backing. Molfino provides you the best Strip Light for your office, home, etc.
What is an Addressable Strp Light And How to Control them?
Whenever we require a dynamic lighting solution, LED strips are an excellent alternative. They are available in a variety of colors, can be mounted almost anywhere, and need little upkeep. You would be hard-pressed to find any lighting system that provides as much value and flair. LED Strip Lights raise the bar to a whole new level. They provide you with more control choices as well as unique effects in lights to make your lighting setup even more effective and brilliant. The catch is that you must select the appropriate type of these strip lights for your requirements. Each LED strip light in a programmable strip lights which includes an inbuilt driver that permits the brightness and color of each result to be set individually; that is, each LED can be a different color than the one next to it. It enables you to apply different effects to different portions of the LED strip at the same time or in intervals. It’s perfect for showing animations and other dynamic displays or for creating beautiful and complicated lighting effects.
Reason for Bright Up Your Space with LED Strip Lights
If you’re getting bored with the old incandescent lighting and you don’t want to go with the on-going CFL or LED bulbs, then LED strip lights are something that you must consider. Given their durability, functionality, colors, safety and decorative uses, the LED flexible strip lights make an excellent choice to be used anywhere. To get more reasons to brighten up your space with LED strip lighting.
Waterproof LED Strip Lights
When we install lights in the kitchen, bathroom or outside, there is a restriction on the type of light fittings you can use. If there is a chance of leakage of water our LED Strip Lights will come into contact with moisture and there is a possibility of short circuit. So, for your safety you will need to use the Strip Lights with a waterproof IP coating. This acts as a protective layer and prevents moisture from reaching the light's internal components, and causing a short circuit. The level of protection offered can vary from unprotected to air tight. If you are using your LED Strip Lights in the kitchen or bathroom you will need a minimum of an IP65 rated coating. The LED chips are encased inside a silicone gel which is able to withstand condensation and splashes of water that may result from cooking food or using taps.
Strip Lights for Home Decoration
LED strip lights have a length of 16.4ft with 90 LED lights. Features remote controller and 12V power adaptor for easy operation. Excellent heat dissipation, extremely low-heat, touchable and safe for children. Strong self-adhesive, can be easily stuck to almost any dry surface, App control LED strip lights.Bring your home theater to life with some help from the smart LED strip lights. The light strip is Wi-Fi enabled, so you can control it with the App or through the included remote control for full customization.LED strip lights is a color LED strip lights kit with a strong adhesive, a control box and wireless remote control. The unique dream color effect, beat following, or music synchronized flashing function creates a romantic and warm atmosphere. Strip lights support 16 million colors to change the color to meet your needs on different occasions.